complete ForumThread; create createForumThread()
NOTE: createForumThread() is buggy now - the content isn't displayed as expected on Guilded client. If u use createForumThread("titlexxx","contentxxx") it will create a topic with titlexxx as title but the content cannot be displayed on computer, and on mobile phone it will directly display "contentxxx"(with quotation marks) as content. I got this problem by following the API doc so it might be a problem on Guilded side